refactor: help page
